Added ecs service class
